+++ monkey/common/event_serializers/__init__.py
@@ -1,5 +1,5 @@
 from .consts import EVENT_TYPE_FIELD
-from .i_event_serializer import IAgentEventSerializer, JSONSerializable
+from .i_agent_event_serializer import IAgentEventSerializer, JSONSerializable
 from .event_serializer_registry import EventSerializerRegistry
 from .pydantic_event_serializer import PydanticEventSerializer
 from .register import register_common_agent_event_serializers
