+++ monkey/common/event_serializers/i_agent_event_serializer.py
